Implement `simplelayout` provider to support multiple simplelayout containers.
