+++ ftw/simplelayout/__init__.py
@@ -1,3 +1,18 @@
 from zope.i18nmessageid import MessageFactory
 
+
 _ = MessageFactory('ftw.simplelayout')
+
+
+def initialize(context):
+
+    from ftw.simplelayout.browser.provider import TalesSimplelayoutExpression
+    from Products.Five.browser.pagetemplatefile import getEngine
+    from zope.tales.tales import RegistrationError
+
+    zope_engine = getEngine()
+    try:
+        zope_engine.registerType('simplelayout', TalesSimplelayoutExpression)
+    except RegistrationError:
+        # zope.reload support
+        pass
